feat(bym): 重构图片处理逻辑，新增 ToDoimg 工具模块
